Redirects Reshuffle to Redistribute, deprecates Reshuffle, merges tests
